FEAT (TextScroll) : adds basic structure
Flicks between two views desktop and mobile

Adds the two views

Adds hook that disctates when the views should change
