+++ src/components/TextScroll/MobileView.tsx
@@ -1,0 +1,23 @@
+import { Box, Flex, Image, Heading } from "@chakra-ui/react";
+import Logo from "../Logo";
+
+export const MobileView = () => (
+  <Box bg="blue" h='80vh'>
+    {/* <Flex w="100%" bg="black" justifyContent="center" alignItems="center">
+      <Image
+        src="./AngleBackground.png"
+        maxH="80vh"
+        minW="100%"
+        h="100%"
+        objectFit="cover"
+        zIndex={2}
+      />
+      <Flex maxW="1500px">
+        <Heading as="h2" color="white">
+          Big Test
+        </Heading>
+        <Logo />
+      </Flex>
+    </Flex> */}
+  </Box>
+);
